# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小知，你的校园助手",
        "title": "KnowEase API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/QA/getbytag": {
            "post": {
                "description": "根据标签获取问答帖子",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "问答"
                ],
                "summary": "根据标签获取问答帖子",
                "parameters": [
                    {
                        "description": "标签名以及用户ID",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功以及问答信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/QA/getgoodqa": {
            "post": {
                "description": "获取精华帖(如果不用标签筛选的话 tag里传一个空字符串)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "问答"
                ],
                "summary": "获取精华帖",
                "parameters": [
                    {
                        "description": "用户ID以及标签信息",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功以及问答信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/QA/{postid}": {
            "get": {
                "description": "获取问答帖子详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "问答"
                ],
                "summary": "获取问答帖子详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功以及问答信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/mypost/deleteqas": {
            "delete": {
                "description": "用户删除已发布的问答",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人主页-我的发布"
                ],
                "summary": "批量删除问答",
                "parameters": [
                    {
                        "description": "帖子id们",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PostIDs"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/mypost/{postid}/deleteqa": {
            "delete": {
                "description": "用户删除已发布的问答",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人主页-我的发布"
                ],
                "summary": "删除问答",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/userpage/{userid}/getuserqa": {
            "get": {
                "description": "获取该用户发布的问答帖子",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "他人主页"
                ],
                "summary": "获取用户问答帖子",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功以及问答信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/QA/publish": {
            "post": {
                "description": "用户发布一个新的问答帖子",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "问答"
                ],
                "summary": "发布问答",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "问答内容",
                        "name": "QA",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QAs"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "发布问答成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/QA/similarity": {
            "post": {
                "description": "根据用户输入的问题，搜索其同质化程度",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "问答"
                ],
                "summary": "搜索相似问答",
                "parameters": [
                    {
                        "description": "用户输入的问题",
                        "name": "QA",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QAs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "未检测到类似问答",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "302": {
                        "description": "检测到有类似问答帖子",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/QA/{postid}/getcounts": {
            "get": {
                "description": "获取帖子的点赞数，收藏数，评论量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子"
                ],
                "summary": "获取问答帖子相关数值",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "相关数值",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Comment": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "commenterID": {
                    "type": "string"
                },
                "commenterName": {
                    "type": "string"
                },
                "commenterURL": {
                    "type": "string"
                },
                "commentid": {
                    "type": "string"
                },
                "creatAt": {
                    "type": "string"
                },
                "imageurl": {
                    "type": "string"
                },
                "likeCount": {
                    "type": "integer"
                },
                "postID": {
                    "type": "string"
                },
                "reply": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Reply"
                    }
                }
            }
        },
...
        "models.PostIDs": {
            "type": "object",
            "properties": {
                "postids": {
                    "type": "string"
                }
            }
        },
...
        "models.QAs": {
            "type": "object",
            "properties": {
                "authorID": {
                    "type": "string"
                },
                "authorName": {
                    "type": "string"
                },
                "authorURL": {
                    "type": "string"
                },
                "comment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Comment"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "likeCount": {
                    "type": "integer"
                },
                "postID": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                },
                "saveCount": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                },
                "updatedAT": {
                    "type": "string"
                },
                "viewCount": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "models.Reply": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "commentID": {
                    "type": "string"
                },
                "commentid": {
                    "type": "string"
                },
                "creatAt": {
                    "type": "string"
                },
                "imageurl": {
                    "type": "string"
                },
                "likeCount": {
                    "type": "integer"
                },
                "postid": {
                    "type": "string"
                },
                "reply": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Reply"
                    }
                },
                "replyURL": {
                    "type": "string"
                },
                "replyerID": {
                    "type": "string"
                },
                "replyerName": {
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Type": {
            "type": "object",
            "properties": {
                "tag": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
    }
}